move remaining sync/async from SepReader, add test to ensure files in sync
